Improve construction of error classes
The error classes encapsulating `errno` were originally converting the
numeric error code to a string representation when the getter for that
string was called, not during construction. This was to mimic virtual
functions so that the right errno-to-string function was called (i.e.
`errno_error` would use `strerror()` and `addrinfo_error` would use
`gai_strerror()`). While it did work, this was inefficient and rather
poor design (in my opinion). This commit remedies that by adding a set
of `protected` constructors to these error classes that except a
function pointer that points to the appropriate errno-to-string
function. The idea is that a more-derived error class can pass an
appropriate function to the base class constructor. The most-base error
class constructor will then convert the `errno` to the appropriate
string representation using the function that has been passed down the
hierarchy through constructors. This public interface remains unchanged.